pyproject.toml
[project]
name = "documentation-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4",
    "black",
    "tiktoken",
    "openai",
    "nltk",
    "fastapi",
    "jinja2",
    "uvicorn",
    "streamlit",
    "streamlit-chat",
    "tqdm",
    "isort",
    "langchainhub",
    "firecrawl-py",
    "langchain-community",
    "langsmith",
    "pillow",
    "certifi",
    "langchain-chroma",
    "python-dotenv",
    "langchain-tavily",
    "langchain-pinecone",
    "protobuf<4.0.0",
    "aiohttp>=3.13.5",
]

[dependency-groups]
dev = [
    "ipykernel",
]

[tool.uv]
required-environments = ["sys_platform == 'darwin' and platform_machine == 'x86_64'"]
